Refactoring the User Form Requests
And fixing the username validation
